refactor(frontend): extract datastudio query execution
