+++ frontend/src/views/datastudio/DataStudioNew.vue
@@ -1147,7 +1147,7 @@
 </template>
 
 <script setup>
-import { defineAsyncComponent, markRaw, nextTick, onBeforeUnmount, onMounted, provide, reactive, ref, watch } from 'vue'
+import { defineAsyncComponent, nextTick, onBeforeUnmount, onMounted, provide, reactive, ref, watch } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 import { ElMessage, ElMessageBox } from 'element-plus'
 import {
@@ -1171,7 +1171,6 @@ import {
 } from '@element-plus/icons-vue'
 import { tableApi } from '@/api/table'
 import { lineageApi } from '@/api/lineage'
-import { dataQueryApi } from '@/api/query'
 import { businessDomainApi, dataDomainApi } from '@/api/domain'
 import PersistentTabs from '@/components/PersistentTabs.vue'
 import TaskEditDrawer from '@/views/tasks/TaskEditDrawer.vue'
@@ -1179,15 +1178,12 @@ import DataStudioResultGrid from '@/views/datastudio/components/DataStudioResult
 import { isDemoMode, showDemoReadonlyMessage } from '@/demo/runtime'
 import { copyText } from '@/utils/clipboard'
 import { loadEcharts } from '@/utils/loadEcharts'
-import { buildCsvContent } from './csvExport'
-import { buildResultGridRows } from './components/resultGridModel'
 import {
   formatNumber,
   formatRowCount,
   formatStorageSize,
   formatDuration,
   formatDateTime,
-  abbreviateSql,
   isAggregateTable,
 } from './tableFormat'
 import {
@@ -1195,13 +1191,13 @@ import {
   scoreMetricColumn,
   detectNumericColumns,
 } from './chartColumnSelect'
-import { splitSqlStatements } from './sqlStatements'
 import { buildFieldPayload, isFieldChanged, isOnlyCommentChanged } from './fieldEdit'
 import { useTabPersistence } from './composables/useTabPersistence'
 import { useResizablePanes } from './composables/useResizablePanes'
 import { useSqlCompletion } from './composables/useSqlCompletion'
 import { useTabRouting } from './composables/useTabRouting'
 import { useCatalogTree } from './composables/useCatalogTree'
+import { useQueryExecution } from './composables/useQueryExecution'
 
 const SqlEditor = defineAsyncComponent({
   loader: () => import('@/components/SqlEditor.vue'),
@@ -1221,9 +1217,6 @@ const DataStudioRightPanel = defineAsyncComponent({
 const clusterId = ref(null)
 const route = useRoute()
 const router = useRouter()
-const historyData = ref([])
-const historyPager = reactive({ pageNum: 1, pageSize: 15, total: 0 })
-const historyLoading = ref(false)
 const createDrawerVisible = ref(false)
 
 const lineageCache = reactive({})
@@ -1232,7 +1225,6 @@ const suppressRouteSync = ref(false)
 const openTabs = ref([])
 const activeTab = ref('')
 const tabStates = reactive({})
-const queryTimerHandles = new Map()
 const queryTabCounter = ref(1)
 const tableObserver = ref(null)
 
@@ -1343,26 +1335,52 @@ const {
   openTableTab: (...args) => openTableTab(...args),
 })
 
+// 查询执行与历史（P2-2 F9）：执行、停止、结果集标准化、导出和历史分页从主组件抽离
+const {
+  historyData,
+  historyPager,
+  historyLoading,
+  buildDefaultSql,
+  clearQueryTimer,
+  stopNowTickerIfIdle,
+  getStatementStatusTagType,
+  isResultSetType,
+  getResultSetCountText,
+  getResultSetAlertType,
+  getDisplayResultSets,
+  handleSqlSelectionChange,
+  handleQuerySourceSelect,
+  handleQueryDatabaseSelect,
+  syncAutoSelectSqlIfSchemaMismatch,
+  executeQuery,
+  stopQuery,
+  getLiveDurationMs,
+  resetQuery,
+  exportResult,
+  fetchHistory,
+  applyHistory,
+  parseResultTabIndex,
+  getResultRowKeyPrefix,
+  getResultSetByIndex,
+} = useQueryExecution({
+  clusterId,
+  activeSource,
+  activeSchema,
+  schemaStore,
+  tabStates,
+  openTabs,
+  activeTab,
+  loadSchemas: (...args) => loadSchemas(...args),
+  loadTables: (...args) => loadTables(...args),
+  syncRouteWithTab: (...args) => syncRouteWithTab(...args),
+  disposeChart: (...args) => disposeChart(...args),
+  applyDefaultChartSelection: (...args) => applyDefaultChartSelection(...args),
+  syncResultPaneLayout: (...args) => syncResultPaneLayout(...args),
+})
+
 	const chartRefs = ref({})
 	const chartInstances = new Map()
 	const taskDrawerRef = ref(null)
-	const nowTick = ref(Date.now())
-	let nowTickHandle = null
-
-	const startNowTicker = () => {
-	  if (nowTickHandle) return
-	  nowTickHandle = setInterval(() => {
-	    nowTick.value = Date.now()
-	  }, 200)
-	}
-
-	const stopNowTickerIfIdle = () => {
-	  if (!nowTickHandle) return
-	  const hasCancelable = Object.values(tabStates).some((state) => !!state?.queryCancelable)
-	  if (hasCancelable) return
-	  clearInterval(nowTickHandle)
-	  nowTickHandle = null
-	}
 
 const layerOptions = [
   { label: 'ODS - 原始数据层', value: 'ODS' },
@@ -1372,32 +1390,6 @@ const layerOptions = [
   { label: 'ADS - 应用数据层', value: 'ADS' }
 ]
 const businessDomainOptions = ref([])
-
-	const clearQueryTimer = (tabId) => {
-	  const handle = queryTimerHandles.get(tabId)
-	  if (!handle) return
-	  clearInterval(handle)
-	  queryTimerHandles.delete(tabId)
-	}
-
-	const startQueryTimer = (tabId) => {
-	  clearQueryTimer(tabId)
-	  const state = tabStates[tabId]
-	  if (!state) return
-	  state.queryTiming.startedAt = Date.now()
-	  state.queryTiming.elapsedMs = 0
-	  startNowTicker()
-	  const handle = setInterval(() => {
-	    const current = tabStates[tabId]
-	    if (!current?.queryCancelable) {
-	      clearQueryTimer(tabId)
-	      stopNowTickerIfIdle()
-	      return
-	    }
-	    current.queryTiming.elapsedMs = Date.now() - current.queryTiming.startedAt
-	  }, 200)
-	  queryTimerHandles.set(tabId, handle)
-	}
 
 const loadBusinessDomains = async () => {
   try {
@@ -1526,93 +1518,6 @@ const warnPlatformMetadataMissing = (table) => {
   if (!isPlatformMetadataMissing(table)) return false
   ElMessage.warning(MISSING_PLATFORM_METADATA_MESSAGE)
   return true
-}
-
-const INFO_TAB_NAME = 'info'
-const RESULT_TYPE_RESULT_SET = 'RESULT_SET'
-const RESULT_TYPE_UPDATE_COUNT = 'UPDATE_COUNT'
-
-const EMPTY_RESULT_SET = Object.freeze({
-  index: 1,
-  statementIndex: 1,
-  status: 'SUCCESS',
-  resultType: RESULT_TYPE_RESULT_SET,
-  affectedRows: null,
-  message: '',
-  sqlSnippet: '',
-  durationMs: 0,
-  columns: [],
-  rows: [],
-  hasMore: false,
-  previewRowCount: 0
-})
-
-const buildRunningStatementInfos = (sqlText) => {
-  const statements = splitSqlStatements(sqlText)
-  return statements.map((statement, idx) => ({
-    statementIndex: idx + 1,
-    status: idx === 0 ? 'RUNNING' : 'PENDING',
-    durationMs: 0,
-    sqlSnippet: abbreviateSql(statement),
-    resultInfo: idx === 0 ? '正在执行' : '等待执行'
-  }))
-}
-
-const buildStatementInfosFromResultSets = (resultSets) => {
-  const sets = Array.isArray(resultSets) ? resultSets : []
-  return sets.map((set, idx) => {
-    const status = String(set?.status || (set?.resultType === 'ERROR' ? 'ERROR' : 'SUCCESS')).toUpperCase()
-    let resultInfo = set?.message || ''
-    if (!resultInfo) {
-      if (String(set?.resultType || '') === RESULT_TYPE_UPDATE_COUNT) {
-        const affected = set?.affectedRows
-        resultInfo = affected === null || affected === undefined ? '语句执行成功' : `影响 ${affected} 行`
-      } else {
-        const rows = Array.isArray(set?.rows) ? set.rows.length : 0
-        resultInfo = `返回 ${rows} 行`
-      }
-    }
-    return {
-      statementIndex: Number(set?.statementIndex || idx + 1),
-      status,
-      durationMs: Number(set?.durationMs || 0),
-      sqlSnippet: set?.sqlSnippet || '',
-      resultInfo
-    }
-  })
-}
-
-const getStatementStatusTagType = (status) => {
-  const value = String(status || '').toUpperCase()
-  if (value === 'SUCCESS') return 'success'
-  if (value === 'RUNNING') return 'info'
-  if (value === 'BLOCKED' || value === 'ERROR') return 'danger'
-  if (value === 'SKIPPED') return 'warning'
-  return 'info'
-}
-
-const isResultSetType = (resultSet) => String(resultSet?.resultType || RESULT_TYPE_RESULT_SET) === RESULT_TYPE_RESULT_SET
-
-const getResultSetCountText = (resultSet) => {
-  const type = String(resultSet?.resultType || RESULT_TYPE_RESULT_SET)
-  if (type === RESULT_TYPE_UPDATE_COUNT) {
-    const affected = resultSet?.affectedRows
-    return affected === null || affected === undefined ? '影响行数未知' : `影响 ${affected} 行`
-  }
-  return `${(resultSet?.rows || []).length} 行`
-}
-
-const getResultSetAlertType = (resultSet) => {
-  const status = String(resultSet?.status || '').toUpperCase()
-  if (status === 'ERROR' || status === 'BLOCKED') return 'error'
-  if (status === 'SKIPPED') return 'warning'
-  return 'success'
-}
-
-const getDisplayResultSets = (tabId) => {
-  const state = tabStates[tabId]
-  const sets = Array.isArray(state?.queryResult?.resultSets) ? state.queryResult.resultSets : []
-  return sets.length ? sets : [EMPTY_RESULT_SET]
 }
 
 const getUpstreamCount = (tableId) => {
@@ -2090,436 +1995,6 @@ const getTabItemById = (tabId) => {
   return openTabs.value.find((tab) => String(tab.id) === String(tabId)) || null
 }
 
-const handleSqlSelectionChange = (tabId, payload) => {
-  const state = tabStates[String(tabId || '')]
-  if (!state) return
-  state.query.selectionText = payload?.text ?? ''
-  state.query.hasSelection = !!payload?.hasSelection
-}
-
-const handleQuerySourceSelect = async (tabId, value) => {
-  const state = tabStates[tabId]
-  const tab = getTabItemById(tabId)
-  if (!state || !tab || tab.kind !== 'query') return
-
-  const sourceId = value ? String(value) : ''
-  state.table.sourceId = sourceId
-  tab.sourceId = sourceId
-
-  state.table.dbName = ''
-  state.table.tableName = ''
-  state.table.id = undefined
-  tab.dbName = ''
-
-  if (String(activeTab.value) === String(tabId)) {
-    clusterId.value = sourceId || null
-    activeSource.value = sourceId
-  }
-
-  if (!sourceId) {
-    if (String(activeTab.value) === String(tabId)) {
-      syncRouteWithTab(tab, tabId)
-    }
-    return
-  }
-
-  const ok = await loadSchemas(sourceId)
-  if (!ok) return
-
-  const nextDb = activeSchema[sourceId] || schemaStore[sourceId]?.[0] || ''
-  if (nextDb) {
-    state.table.dbName = nextDb
-    tab.dbName = nextDb
-    activeSchema[sourceId] = nextDb
-    await loadTables(sourceId, nextDb)
-  }
-
-  if (String(activeTab.value) === String(tabId)) {
-    syncRouteWithTab(tab, tabId)
-  }
-}
-
-const handleQueryDatabaseSelect = async (tabId, value) => {
-  const state = tabStates[tabId]
-  const tab = getTabItemById(tabId)
-  if (!state || !tab || tab.kind !== 'query') return
-
-  const dbName = value ? String(value) : ''
-  state.table.dbName = dbName
-  tab.dbName = dbName
-
-  state.table.tableName = ''
-  state.table.id = undefined
-
-  const sourceId = String(state.table.sourceId || tab.sourceId || '')
-  if (sourceId && dbName) {
-    activeSchema[sourceId] = dbName
-    await loadTables(sourceId, dbName)
-  }
-
-  if (String(activeTab.value) === String(tabId)) {
-    clusterId.value = sourceId || null
-    activeSource.value = sourceId
-    syncRouteWithTab(tab, tabId)
-  }
-}
-
-const buildDefaultSql = (table) => {
-  if (!table?.dbName || !table?.tableName) return ''
-  return `SELECT *\nFROM \`${table.dbName}\`.\`${table.tableName}\`\nLIMIT 200;`
-}
-
-const parseAutoSelectSql = (sql) => {
-  const text = String(sql || '').trim()
-  if (!text) return null
-  const match = text.match(/^select\s+\*\s+from\s+`([^`]+)`\.`([^`]+)`\s+limit\s+(\d+)\s*;?$/i)
-  if (!match) return null
-  return { schema: match[1], table: match[2], limit: Number(match[3]) }
-}
-
-const syncAutoSelectSqlIfSchemaMismatch = (state) => {
-  if (!state?.table?.dbName || !state?.table?.tableName) return
-  const nextDefault = buildDefaultSql(state.table)
-  if (!String(state.query?.sql || '').trim()) {
-    state.query.sql = nextDefault
-    return
-  }
-  const parsed = parseAutoSelectSql(state.query.sql)
-  if (!parsed) return
-  if (parsed.table === state.table.tableName && parsed.schema !== state.table.dbName) {
-    state.query.sql = nextDefault
-  }
-}
-
-	const executeQuery = async (tabId) => {
-	  const state = tabStates[tabId]
-	  if (!state) return
-	  const runId = Number(state.queryRunId || 0) + 1
-	  state.queryRunId = runId
-	  const selectedSql = String(state?.query?.selectionText || '')
-	  const sqlToRun = selectedSql.trim() ? selectedSql : String(state?.query?.sql || '')
-  if (!sqlToRun.trim()) {
-    state.queryResult.errorMessage = '请输入 SQL'
-    state.queryResult.message = ''
-    state.resultTab = INFO_TAB_NAME
-    return
-  }
-  if (!state.table?.dbName) {
-    state.queryResult.errorMessage = '请先选择数据库'
-    state.queryResult.message = ''
-    state.resultTab = INFO_TAB_NAME
-    return
-  }
-  const sourceId = state.table?.sourceId || clusterId.value
-	  if (!sourceId) {
-	    state.queryResult.errorMessage = '请选择数据源'
-	    state.queryResult.message = ''
-	    state.resultTab = INFO_TAB_NAME
-	    return
-	  }
-
-  let analyzeRes = null
-  try {
-    analyzeRes = await dataQueryApi.analyze({
-      clientQueryId: String(tabId),
-      clusterId: sourceId || undefined,
-      database: state.table.dbName || undefined,
-      sql: sqlToRun
-    })
-  } catch (error) {
-    const message = error?.response?.data?.message || error?.message || 'SQL 分析失败'
-    state.queryResult = {
-      resultSets: [],
-      columns: [],
-      rows: [],
-      hasMore: false,
-      durationMs: 0,
-      executedAt: '',
-      cancelled: false,
-      statementInfos: [
-        {
-          statementIndex: 1,
-          status: 'ERROR',
-          durationMs: 0,
-          sqlSnippet: abbreviateSql(sqlToRun),
-          resultInfo: message
-        }
-      ],
-      message: '',
-      errorMessage: message
-    }
-    state.resultTab = INFO_TAB_NAME
-    return
-  }
-
-  const blockedRiskItem = Array.isArray(analyzeRes?.riskItems)
-    ? analyzeRes.riskItems.find((item) => item?.blocked)
-    : null
-  const blockedStatementIndex = Number(blockedRiskItem?.statementIndex || 0) || null
-  const confirmChallenges = Array.isArray(analyzeRes?.confirmChallenges)
-    ? [...analyzeRes.confirmChallenges]
-      .filter((item) => {
-        const idx = Number(item?.statementIndex || 0)
-        return !blockedStatementIndex || (idx > 0 && idx < blockedStatementIndex)
-      })
-      .sort((a, b) => Number(a?.statementIndex || 0) - Number(b?.statementIndex || 0))
-    : []
-  const confirmations = []
-  for (const challenge of confirmChallenges) {
-    const expected = String(challenge?.targetObject || '').trim()
-    try {
-      const { value } = await ElMessageBox.prompt(
-        `语句 #${challenge.statementIndex} 为高风险操作，请输入对象名确认执行：${expected}`,
-        '高风险 SQL 强确认',
-        {
-          type: 'warning',
-          confirmButtonText: '确认执行',
-          cancelButtonText: '取消',
-          inputValue: '',
-          inputPlaceholder: expected,
-          inputValidator: (input) => {
-            if (String(input || '').trim() !== expected) {
-              return `请输入对象名：${expected}`
-            }
-            return true
-          }
-        }
-      )
-      confirmations.push({
-        statementIndex: Number(challenge?.statementIndex || 0),
-        targetObject: expected,
-        inputText: String(value || '').trim(),
-        confirmToken: challenge?.confirmToken || ''
-      })
-    } catch (error) {
-      if (error === 'cancel' || error === 'close') {
-        break
-      }
-      const message = error?.response?.data?.message || error?.message || '强确认失败'
-      state.queryResult = {
-        resultSets: [],
-        columns: [],
-        rows: [],
-        hasMore: false,
-        durationMs: 0,
-        executedAt: '',
-        cancelled: false,
-        statementInfos: [
-          {
-            statementIndex: Number(challenge?.statementIndex || 1),
-            status: 'ERROR',
-            durationMs: 0,
-            sqlSnippet: challenge?.targetObject || abbreviateSql(sqlToRun),
-            resultInfo: message
-          }
-        ],
-        message: '',
-        errorMessage: message
-      }
-      state.resultTab = INFO_TAB_NAME
-      return
-    }
-  }
-
-		  if (state.queryAbortController) {
-	    try {
-	      state.queryAbortController.abort()
-	    } catch (_) {
-	      // ignored
-	    }
-	  }
-	  state.queryAbortController = new AbortController()
-	  state.queryLoading = true
-	  state.queryStopping = false
-	  state.queryCancelable = true
-	  startNowTicker()
-	  state.queryResult.errorMessage = ''
-	  state.queryResult.message = ''
-	  state.queryResult.cancelled = false
-	  state.queryResult.statementInfos = buildRunningStatementInfos(sqlToRun)
-	  state.resultTab = INFO_TAB_NAME
-	  startQueryTimer(tabId)
-	  disposeChart(tabId)
-	  try {
-	    const res = await dataQueryApi.execute({
-	      clientQueryId: String(tabId),
-	      clusterId: sourceId || undefined,
-	      database: state.table.dbName || undefined,
-	      sql: sqlToRun,
-	      limit: state.query.limit,
-	      confirmations
-	    }, { signal: state.queryAbortController?.signal })
-	    if (state.queryRunId !== runId) return
-
-	    const resultSets = Array.isArray(res.resultSets) ? res.resultSets : []
-    const fallbackResultSet = {
-      index: 1,
-      statementIndex: 1,
-      status: 'SUCCESS',
-      resultType: RESULT_TYPE_RESULT_SET,
-      affectedRows: null,
-      message: res.message || '',
-      sqlSnippet: abbreviateSql(sqlToRun),
-      durationMs: Number(res.durationMs || 0),
-      columns: res.columns || [],
-      rows: res.rows || [],
-      hasMore: !!res.hasMore,
-      previewRowCount: (res.rows || []).length
-    }
-    const normalizedSets = normalizeResultSetsForDisplay(resultSets.length ? resultSets : [fallbackResultSet], tabId)
-    const statementInfos = buildStatementInfosFromResultSets(normalizedSets)
-    const hasFailure = normalizedSets.some((item) => {
-      const status = String(item?.status || '').toUpperCase()
-      return status === 'BLOCKED' || status === 'ERROR' || status === 'SKIPPED'
-    })
-
-		    state.queryResult = {
-		      resultSets: normalizedSets,
-		      columns: normalizedSets[0]?.columns || [],
-		      rows: normalizedSets[0]?.rows || [],
-      hasMore: res.hasMore,
-      durationMs: res.durationMs,
-      executedAt: res.executedAt,
-      cancelled: !!res.cancelled,
-		      statementInfos,
-		      message: res.message || '',
-		      errorMessage: ''
-		    }
-		    state.queryCancelable = false
-		    state.queryAbortController = null
-		    stopNowTickerIfIdle()
-		    state.page.current = 1
-		    state.resultTab = !res.cancelled && !hasFailure ? 'result-0' : INFO_TAB_NAME
-			    state.charts = normalizedSets.map(() => ({
-			      type: 'bar',
-		      xAxis: '',
-		      yAxis: []
-		    }))
-			    state.resultViewTabs = normalizedSets.map((_, idx) => state.resultViewTabs?.[idx] || 'table')
-			    applyDefaultChartSelection(tabId)
-			    await nextTick()
-			    syncResultPaneLayout(tabId)
-			    fetchHistory()
-		  } catch (error) {
-		    if (state.queryRunId !== runId) return
-		    const isCanceled =
-		      String(error?.code || '') === 'ERR_CANCELED' ||
-		      String(error?.name || '') === 'CanceledError' ||
-		      /canceled/i.test(String(error?.message || ''))
-		    if (isCanceled) {
-		      return
-		    }
-		    const message = error?.response?.data?.message || error?.message || '查询失败'
-		    const hasResponse = !!error?.response
-		    const maybeStillRunning = !hasResponse
-		    if (!maybeStillRunning) {
-		      state.queryCancelable = false
-		    }
-		    if (!state.queryCancelable) {
-		      state.queryAbortController = null
-		      stopNowTickerIfIdle()
-		    }
-		    state.queryResult = {
-	      resultSets: [],
-	      columns: [],
-	      rows: [],
-      hasMore: false,
-	      durationMs: 0,
-	      executedAt: '',
-	      cancelled: false,
-		      statementInfos: maybeStillRunning
-		        ? (Array.isArray(state.queryResult?.statementInfos) ? state.queryResult.statementInfos : buildRunningStatementInfos(sqlToRun))
-		        : [
-		          {
-		            statementIndex: 1,
-		            status: 'ERROR',
-		            durationMs: 0,
-		            sqlSnippet: abbreviateSql(sqlToRun),
-		            resultInfo: message
-		          }
-		        ],
-		      message: maybeStillRunning ? '查询请求超时/网络异常，可能仍在执行，可点击“停止”' : '',
-		      errorMessage: message
-		    }
-			    state.resultTab = INFO_TAB_NAME
-		    state.charts = [
-		      {
-		        type: 'bar',
-		        xAxis: '',
-		        yAxis: []
-		      }
-			    ]
-			    state.resultViewTabs = ['table']
-				  } finally {
-			    if (state.queryRunId !== runId) return
-			    state.queryLoading = false
-			    if (!state.queryCancelable) {
-			      clearQueryTimer(tabId)
-			    }
-			  }
-	}
-
-	const stopQuery = async (tabId) => {
-	  const state = tabStates[tabId]
-	  if (!state?.queryCancelable || state.queryStopping) return
-	  state.queryStopping = true
-	  try {
-	    state.queryAbortController?.abort()
-	  } catch (_) {
-	    // ignored
-	  }
-	  state.queryAbortController = null
-	  try {
-	    await dataQueryApi.stop({ clientQueryId: String(tabId) })
-	    state.queryCancelable = false
-	    state.queryLoading = false
-	    state.queryStopping = false
-	    clearQueryTimer(tabId)
-	    stopNowTickerIfIdle()
-	    state.queryResult.cancelled = true
-	    state.queryResult.message = '查询已停止'
-	    state.queryResult.errorMessage = ''
-	    const existingInfos = Array.isArray(state.queryResult.statementInfos) ? state.queryResult.statementInfos : []
-	    state.queryResult.statementInfos = existingInfos.map((item, idx) => {
-	      const status = String(item?.status || '').toUpperCase()
-	      if (status === 'SUCCESS' || status === 'ERROR' || status === 'BLOCKED') return item
-	      return {
-	        statementIndex: Number(item?.statementIndex || idx + 1),
-	        status: 'SKIPPED',
-	        durationMs: Number(item?.durationMs || 0),
-	        sqlSnippet: item?.sqlSnippet || '',
-	        resultInfo: '查询已停止'
-	      }
-	    })
-	    state.resultTab = INFO_TAB_NAME
-	  } catch (error) {
-	    state.queryStopping = false
-	    const message = error?.response?.data?.message || error?.message || '停止失败'
-	    state.queryResult.errorMessage = message
-	    state.queryResult.message = ''
-	    state.resultTab = INFO_TAB_NAME
-	  }
-	}
-
-	const getLiveDurationMs = (tabId) => {
-	  const state = tabStates[String(tabId || '')]
-	  if (!state) return 0
-	  if (state.queryCancelable) {
-	    const startedAt = Number(state.queryTiming?.startedAt || 0)
-	    if (!Number.isFinite(startedAt) || startedAt <= 0) return 0
-	    return Math.max(0, nowTick.value - startedAt)
-	  }
-	  return Number(state.queryResult?.durationMs || 0)
-	}
-
-const resetQuery = (tabId) => {
-  const state = tabStates[tabId]
-  if (!state) return
-  state.query.sql = buildDefaultSql(state.table)
-  state.query.selectionText = ''
-  state.query.hasSelection = false
-}
-
 const saveAsTask = (tabId) => {
   if (isDemoMode) {
     showDemoReadonlyMessage('保存查询任务')
@@ -2536,58 +2011,6 @@ const saveAsTask = (tabId) => {
     taskName: `新建查询任务_${Date.now()}`,
     taskDesc: `From DataStudio\nCluster: ${sourceId}\nDatabase: ${state.table.dbName || ''}`
   })
-}
-
-const exportResult = (tabId, resultIndex = 0) => {
-  const state = tabStates[tabId]
-  if (!state?.queryResult) return
-  const idx = Number(resultIndex)
-  const set = Array.isArray(state.queryResult.resultSets) ? state.queryResult.resultSets[idx] : null
-  const columns = set?.columns || state.queryResult.columns || []
-  const rows = set?.rows || state.queryResult.rows || []
-  if (!rows.length || !columns.length) return
-
-  const blob = new Blob([buildCsvContent(columns, rows)], { type: 'text/csv;charset=utf-8;' })
-  const link = document.createElement('a')
-  link.href = URL.createObjectURL(blob)
-  link.download = `export_${Date.now()}.csv`
-  link.click()
-}
-
-const fetchHistory = async () => {
-  historyLoading.value = true
-  try {
-    const res = await dataQueryApi.history({
-      pageNum: historyPager.pageNum,
-      pageSize: historyPager.pageSize
-    })
-    historyData.value = res.records || []
-    historyPager.total = res.total || 0
-  } catch (error) {
-    console.error('加载历史查询失败', error)
-  } finally {
-    historyLoading.value = false
-  }
-}
-
-const applyHistory = (row, tabId) => {
-  const state = tabStates[tabId]
-  if (!state || !row) return
-  state.query.sql = row.sqlText || ''
-  if (row.clusterId) {
-    const sourceId = String(row.clusterId)
-    clusterId.value = row.clusterId
-    activeSource.value = sourceId
-    loadSchemas(row.clusterId)
-    const tab = getTabItemById(tabId)
-    if (tab?.kind === 'query') {
-      tab.sourceId = sourceId
-      state.table.sourceId = sourceId
-      if (String(activeTab.value) === String(tabId)) {
-        syncRouteWithTab(tab, tabId)
-      }
-    }
-  }
 }
 
 const handleCreateTable = () => {
@@ -2753,52 +2176,7 @@ const handleCreateSuccess = async (result) => {
   }
 }
 
-
-const getPaginatedRows = (tabId) => {
-  const state = tabStates[tabId]
-  if (!state) return []
-  const start = (state.page.current - 1) * state.page.size
-  const end = start + state.page.size
-  return state.queryResult.rows.slice(start, end)
-}
-
-const parseResultTabIndex = (value) => {
-  const match = String(value || '').match(/^result-(\d+)$/)
-  return match ? Number(match[1]) : null
-}
-
 const getChartKey = (tabId, resultIndex) => `${String(tabId)}::${Number(resultIndex)}`
-const getResultRowKeyPrefix = (tabId, resultIndex) => `${String(tabId)}::${Number(resultIndex)}`
-
-const normalizeResultSetForDisplay = (resultSet, tabId, resultIndex) => {
-  const rows = Array.isArray(resultSet?.rows) ? resultSet.rows : []
-  const columns = Array.isArray(resultSet?.columns) ? resultSet.columns : []
-  return markRaw({
-    ...resultSet,
-    columns,
-    rows: markRaw(buildResultGridRows(rows, getResultRowKeyPrefix(tabId, resultIndex))),
-    hasMore: !!resultSet?.hasMore,
-    previewRowCount: Number.isFinite(Number(resultSet?.previewRowCount))
-      ? Number(resultSet.previewRowCount)
-      : rows.length
-  })
-}
-
-const normalizeResultSetsForDisplay = (resultSets, tabId) => {
-  const sets = Array.isArray(resultSets) ? resultSets : []
-  return markRaw(sets.map((set, idx) => normalizeResultSetForDisplay(set, tabId, idx)))
-}
-
-	const getResultSetByIndex = (tabId, resultIndex = 0) => {
-	  const state = tabStates[tabId]
-  const sets = Array.isArray(state?.queryResult?.resultSets) ? state.queryResult.resultSets : []
-  const set = sets[resultIndex] || EMPTY_RESULT_SET
-  return {
-    columns: Array.isArray(set?.columns) ? set.columns : [],
-    rows: Array.isArray(set?.rows) ? set.rows : [],
-    hasMore: !!set?.hasMore
-  }
-}
 
 const getNumericColumns = (tabId, resultIndex = 0) => {
   const set = getResultSetByIndex(tabId, resultIndex)
@@ -3392,13 +2770,6 @@ const goLineage = (tabId) => {
 }
 
 watch(
-  () => [historyPager.pageNum, historyPager.pageSize],
-  () => {
-    fetchHistory()
-  }
-)
-
-watch(
   () => {
 	    const tabId = activeTab.value
 	    if (!tabId) return null
@@ -3556,10 +2927,8 @@ onMounted(async () => {
 
 onBeforeUnmount(() => {
   flushPersistTabs()
-				chartInstances.forEach((instance) => instance.dispose())
-				chartInstances.clear()
-			queryTimerHandles.forEach((handle) => clearInterval(handle))
-			queryTimerHandles.clear()
+					chartInstances.forEach((instance) => instance.dispose())
+					chartInstances.clear()
   if (tableObserver.value) {
     tableObserver.value.disconnect()
   }
